// frontend/src/app/model/Acte.ts
export class Acte{
    act_numero: Number;
    act_numeropolice: Number;
    act_typegarantie: Number;
    act_typemarche: String;
    act_numeromarchepublic: String;
    act_codemarche: Number;
    act_datemarche: Date;
    act_idcontractante: Number;
    act_idbeneficiaire: Number;
    act_idcandidat: Number;
    act_descriptionmarche;
    act_capitalassure: Number;
    act_capitalsmp: Number;
    act_capitallci: Number;
    act_anciennumero;
    act_status: String;
    act_datecomptabilisation: Date;
}