// frontend/src/app/model/Avenant.ts
export class Avenant{
    aven_id: Number;
    aven_numeropolice: Number;
    aven_numeroavenant: Number;
    aven_codeintermediare: Number;
    aven_codecompagnie: String;
    aven_codemarche: Number;
    aven_typavenant: Number;
    aven_dateeffet: Date;
    aven_dateecheance: Date;
    aven_duree: Number;
    aven_daterevision: Date;
    aven_typerevision: Number;
    aven_typecontrat: Number;
    aven_typegestion: Number;
    aven_typefraction: Number;
    aven_montantprimenet: Number;
    aven_montantprimebrut: Number;
    aven_coefbonus: Number;
    aven_coefremisecie: Number;
    aven_numeroattestionauto: String;
    aven_numerocertificat: String;
    aven_numpoliceremplacee: Number;
    aven_typefacultative: String;
    aven_numeroplancoassur: Number;
    aven_numeroplanreassur: Number;
    aven_numeroderniernavenant: Number;
    aven_datesuspension: Date;
    aven_dateremisevigueur: Date;
    aven_dateresiliation: Date;
    aven_statut: String;
    aven_indice: Number;
    aven_typeindice: String;
    aven_dateindice: Date;
    aven_valeurindice: Number;
    aven_exonerationtaxe: String;
    aven_codetaxe: Number;
    aven_exonerationtva
    aven_codetva: Number;
    aven_exonerationtps: String;
    aven_codetps: Number;
    aven_dateexoneration: Date;
    aven_formulegarantie: Number;
    aven_utilisateur: Number;
    aven_datemodification: Date;
}