// frontend/src/app/model/Marche.ts
export class Marche{
    march_numero: Number;
	march_identification: String;
	march_idcontractante: Number;
	march_donneurordre: Number;
	march_descriptionmarche1: String;
	march_descriptionmarche2: String;
	march_nombrelots: Number;
	march_dureeenmois: Number;
	march_capitalglobalmarche: Number;
	march_capitalglobalsmpmarche: Number;
	march_capitalglobalslcimarche: Number;
	march_status: Number;
	march_datemodification: Date;
	march_codeutil: String;
}